requirements.txt
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
gTTS==2.2.4
idna==3.7
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pillow==10.4.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.19
Werkzeug==2.0.3
zipp==3.6.0
